Tweak "bright black" text color
This improves the appearance of a few programs (mocha, Node.js REPL
displaying `undefined` values, etc.)

This color was #073642 (same as solarized dark background).  This commit
changes it to #083642 because Terminal.app applies some kind of filter
to the colors, and this way the result of that filter ends up lighter
than the original color rather than close to pure black.

Also tried blending between #073642 ("black") and #586e75 ("bright
green") with the http://meyerweb.com/eric/tools/color-blend/ tool.  This
did not yield any better results: it's possible to make "bright black on
black background" visible this way but why would a program ever use that
combination??